Add message alert on all actions

+++ server/childTables.php
@@ -1,8 +1,8 @@
 <?php
 session_start();
-if(!isset($_SESSION['login']) && $_SESSION['user']=="member"){
+if (!isset($_SESSION['login']) && $_SESSION['user'] == "member") {
     header("Location: ../client/member_login.php");
-}else if(!isset($_SESSION['login']) && $_SESSION['user']=="admin"){
+} else if (!isset($_SESSION['login']) && $_SESSION['user'] == "admin") {
     header("Location: ../client/admin_login.php");
 }
 include "../db/conn.php";
@@ -17,12 +17,19 @@ include "../includes/sidebar.php";
             <div class="card " style="width: 1250px;">
                 <div class="card-header">
 
-                    <h3 class="card-title"> All Children </h3>
+                    <h3 class="card-title"> All Children
+                        <?php if (isset($_GET['msg'])) { ?>
+                            <span id="msg" class="alert alert-success" role="alert">
+                                <i class="fa fa-check-circle"></i>
+                                <?= $_GET['msg'] ?>
+                            </span>
+                        <?php }
+                        ?>
+                    </h3>
 
                     <div class="card-tools">
                         <div class="input-group input-group-sm">
                             <input type="text" name="table_search" class="form-control float-right" placeholder="Search">
-
                             <div class="input-group-append">
                                 <button type="submit" class="btn btn-default">
                                     <i class="fas fa-search"></i>
@@ -48,7 +55,8 @@ include "../includes/sidebar.php";
                         <tbody>
                             <?php
                             if (isset($_GET['filter']) && $_GET['filter'] == 'marriageable') {
-                                $sql = "SELECT * FROM child WHERE isMarriageable='yes'";
+                                //change it later
+                                $sql = "SELECT * FROM child WHERE gender='Male'";
                             } else {
                                 $sql = "SELECT * FROM child";
                             }
@@ -67,14 +75,14 @@ include "../includes/sidebar.php";
                                         <td><?= $row["age"] ?></td>
                                         <td><button type="submit">Active</button></td>
                                         <td>
-                                            
+
                                             <button class="btn btn-primary ml-1" data-toggle="modal" data-target="#example"><i class="fas fa-eye"></i>
                                             </button>
                                             <a href="./childUpdate.php?id=<?= $row['id'] ?> "><button class="btn btn-primary ml-1"><i class="fas fa-edit"></i></button></a>
-                                            <a href="../process/deleteChild.php?id=<?= $row['id'] ?>"><button type="submit" class="btn btn-danger ml-1" id="delete"><i class="fas fa-trash-alt"></i></button></a>
+                                            <a href="../process/deleteChild.php?id=<?= $row['id'] ?>" onclick="return confirm('Are you sure?')"><button type="submit" class="btn btn-danger ml-1" id="delete"><i class="fas fa-trash-alt"></i></button></a>
                                         </td>
                                     </tr>
-                                    <?php include "../includes/modal.php";?>
+                                    <?php include "../includes/modal.php"; ?>
                             <?php
                                 }
                             } else {
